ドラッグ操作によるランクの並び替えを実装 2015-07-06  Hidehisa Akiyama  <akky@users.sourceforge.jp>
	* src/qt4/action_sequence_selector.cpp:
	* src/qt4/action_sequence_selector.h:
	* src/qt4/action_sequence_tree_view.cpp:
	* src/qt4/action_sequence_tree_view.h:
	- ドラッグ操作によるランクの並び替えを実装。評価値は上下２つの平均
	値に設定される。端の場合は+1または-1。

+++ src/qt4/action_sequence_tree_view.cpp
@@ -37,13 +37,17 @@
 
 #include "action_sequence_tree_view.h"
 
+#include <iostream>
+#include <limits>
+
 
 /*-------------------------------------------------------------------*/
 /*!
 
 */
 ActionSequenceTreeView::ActionSequenceTreeView( QWidget * parent )
-    : QTreeWidget( parent )
+    : QTreeWidget( parent ),
+      M_pressed_item( static_cast< QTreeWidgetItem * >( 0 ) )
 {
     //this->setContextMenuPolicy( Qt::CustomContextMenu );
     this->setSelectionBehavior( QAbstractItemView::SelectRows );
@@ -52,14 +56,14 @@ ActionSequenceTreeView::ActionSequenceTreeView( QWidget * parent )
     this->setAlternatingRowColors( true );
     this->setAutoScroll( true );
     this->setRootIsDecorated( false );
-#if 0
+#if 1
+    // settings for drag and drop reordering
     this->setDragEnabled( true );
     this->setDropIndicatorShown( true );
     this->setDragDropMode( QAbstractItemView::InternalMove );
     this->viewport()->setAcceptDrops( true );
 #endif
     this->setEditTriggers( QAbstractItemView::NoEditTriggers ); // handled only by double click
-    //this->setEditTriggers( QAbstractItemView::SelectedClicked );
 
     {
         QTreeWidgetItem * h = this->headerItem();
@@ -79,6 +83,9 @@ ActionSequenceTreeView::ActionSequenceTreeView( QWidget * parent )
     this->setColumnWidth( VALUE_COLUMN, metrics.width( tr( "000000.000" ) ) );
     this->setColumnWidth( LENGTH_COLUMN, metrics.width( tr( "00" ) ) );
     this->setColumnWidth( SEQ_COLUMN, metrics.width( tr( "0000" ) ) );
+
+    connect( this, SIGNAL( itemPressed( QTreeWidgetItem *, int ) ),
+             this, SLOT( slotItemPressed( QTreeWidgetItem *, int ) ) );
 }
 
 /*-------------------------------------------------------------------*/
@@ -97,13 +104,18 @@ ActionSequenceTreeView::~ActionSequenceTreeView()
 void
 ActionSequenceTreeView::dropEvent( QDropEvent * event )
 {
-#if 1
-    QTreeWidget::dropEvent( event );
-#else
-    QModelIndex drop_index = this->indexAt( event->pos() );
-
-    if ( drop_index.isValid() == false)
+    if ( this->sortColumn() != VALUE_COLUMN )
     {
+        std::cerr << "(dropEvent) sortColumn != VALUE" << std::endl;
+        event->setDropAction( Qt::IgnoreAction );
+        event->accept();
+        return;
+    }
+
+    QModelIndex drop_index = this->indexAt( event->pos() );
+    if ( ! drop_index.isValid() )
+    {
+        std::cerr << "(dropEvent) invalid drop_index" << std::endl;
         event->setDropAction( Qt::IgnoreAction );
         event->accept();
         return;
@@ -111,11 +123,83 @@ ActionSequenceTreeView::dropEvent( QDropEvent * event )
 
     QTreeWidget::dropEvent( event );
 
-    DropIndicatorPosition drop_pos = this->dropIndicatorPosition();
-    if ( dp == QAbstractItemView::BelowItem )
+    // std::cerr << "(dropEvent) drop_index=(" << drop_index.row() << ","
+    //           << drop_index.column() << ")" << std::endl;
+
+    if ( ! M_pressed_item )
     {
-        drop_index = drop_index.sibling(droppedIndex.row() + 1, // adjust the row number
-                                            droppedIndex.column());
+        std::cerr << "(dropEvent) no pressed item" << std::endl;
+        return;
     }
-#endif
+
+    DropIndicatorPosition dip = this->dropIndicatorPosition();
+
+    QTreeWidgetItem * prev = static_cast< QTreeWidgetItem * >( 0 );
+    QTreeWidgetItem * next = static_cast< QTreeWidgetItem * >( 0 );
+
+    if ( dip == QAbstractItemView::AboveItem )
+    {
+        if ( drop_index.row() > 0 )
+        {
+            prev = this->topLevelItem( drop_index.row() - 1 );
+        }
+        next = this->topLevelItem( drop_index.row() );
+    }
+    else if ( dip == QAbstractItemView::BelowItem )
+    {
+        prev = this->topLevelItem( drop_index.row() );
+        if ( drop_index.row() < this->topLevelItemCount() )
+        {
+            next = this->topLevelItem( drop_index.row() + 1 );
+        }
+    }
+
+    if ( ! prev
+         && ! next )
+    {
+        std::cerr << "" << std::endl;
+        return;
+    }
+
+    double old_value = M_pressed_item->data( VALUE_COLUMN, Qt::DisplayRole ).toDouble();
+    double new_value = old_value;
+
+    if ( prev
+         && ! next )
+    {
+        double prev_value = prev->data( VALUE_COLUMN, Qt::DisplayRole ).toDouble();
+        new_value = prev_value - 1.0;
+    }
+
+    if ( ! prev
+         && next )
+    {
+        double next_value = next->data( VALUE_COLUMN, Qt::DisplayRole ).toDouble();
+        new_value = next_value + 1.0;
+    }
+
+    if ( prev
+         && next )
+    {
+        double prev_value = prev->data( VALUE_COLUMN, Qt::DisplayRole ).toDouble();
+        double next_value = next->data( VALUE_COLUMN, Qt::DisplayRole ).toDouble();
+        new_value = ( prev_value + next_value ) * 0.5;
+    }
+
+    std::cerr << "update value " << old_value << " -> " << new_value
+              << std::endl;
+    M_pressed_item->setData( VALUE_COLUMN, Qt::DisplayRole, new_value );
+
+    emit itemModified();
+}
+
+/*-------------------------------------------------------------------*/
+/*!
+
+*/
+void
+ActionSequenceTreeView::slotItemPressed( QTreeWidgetItem * item,
+                                         int /*column*/ )
+{
+    M_pressed_item = item;
 }
